Allow multiple inline SVGs per page
If more than one SVG is placed inline in an HTML file, the latter SVGs
render not only their own animations, but the animations of the SVGs
above them.

This patch makes sure that all the IDs used in an SVG are different from
those in another SVG so that the animations don't bleed into each other.

Also, the CSS applied to the SVGs is limited in scope to just SVGs